requirements/dev.txt
confluent_kafka==2.5.0
# pyspark~=3.3.1
# pandas~=1.5.3
pillow==10.3.0
opencv-python==4.9.0.80
numpy==1.26.3
minio==7.2.5
